pyproject.toml
[project]
name = "folio-benchmark"
version = "1.0.0"
description = "FOLIO Logical Reasoning Benchmark for AgentBeats"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk[http-server]>=0.3.8",
    "litellm>=1.0.0",
    "pandas>=2.0.0",
    "typer>=0.9.0",
    "uvicorn>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []

